Fix SpiralResonatorPolygon last segment problem
The original bug was presented in https://gitlab.iqm.fi/iqm/qe/KQCircuits/-/issues/768 . Reason for the bug was that _produce_resonator allowed too short last segment and _fix_waveguide_end was not able to handle it.

This commit also changes the method to avoid last segment overlap with an outer segment. This simplifies the code in _fix_waveguide_end. Now, resonator ends more often with straight segment than in the previous version.

Since the nanometer scale error problem is solved, there is no need for using corner_safety_overlap. So this parameters is removed.