Speed up SpiralResonatorPolygon
- Checking, if the resonator spacing is ok for creating a resonator, is now located in function `_produce_path_points`. This function doesn't use WaveguideCoplanar elements, which makes iteration much quicker than it was before.
- The `_produce_resonator` function takes points generated in `_produce_path_points` as an argument.
- Parameter `spacing_tolerance` is removed and hardcoded value 0.001 is used instead, because there is no significant speed advantage by having larger tolerance.

There is a small chance that the new implementation has slightly different geometry than the previous implementation. Difference can occur on the limit of deciding whether the last segment is straight or curve (in those cases both options are ok, so no harm is done). In most cases the resonator geometry is exactly the same as in the previous implementation.