Matthias' workaround
"KLayout evaluates the PCell generation code in the context of the library, not
the target layout." When a PCell uses an other PCell from the same library it
should not use explicit library name in `layout.create_cell()` as this is
automatically detected by KLayout.

Also, this will not show extra library names in the instantiation path of
composite PCells in the cell tree.