Add junction-widths parameter to JunctionTestPads
junction-widths is an optional list of SQUID junction width values that are
sequentially used in the generated SQUIDs. If there are not enough elements in
the list the rest of the SQUIDs will use the default junction-width value.

Alternatively the junction_width_steps (start, step) parameter may be used to auto-generate all
the SQUID junction width values.

Issue #736