Replace a specific SQUID in a design file
by a new SQUID of different type, mirroring or other parameters. The SQUID is
specified by the name of its containing Qubit or Test Structure. If there are
more than one SQUIDs in the specified container squid_index is used to select
which one to replace.

All SQUIDs identified by the given name (and squid_index) will be replaced if
the design file contains several elements with identical names.

Issue #843