The gap_width parameter now determines only the width of the gap
Previously gap_width determined the width of the
entire Swissmon arm, including the gap and the island.
Now, for each four arms in Swissmon, the width of the
gap is parametrised with gap_width and the width of
the island is parametrised with arm_width.
Additionally, refactored code so not to confuse west and east
for ww and we variables.