Use simulation gap for mesh refinement in Ansys
This MR implements the "simulation gap" layers to Ansys.

The layers are then used for restricting the element length in
the gap with `gap_max_element_length` parameter. This will give a
better initial mesh that is helpful for example in eigenmode simulations.

The feature is active only if gap_max_element_length is given in the
ansys export parameters.

A test case is added:
  * klayout_package/python/scripts/simulations/waveguide_eig_mesh_test.py

Issue #650