Generalize face naming system
The old 'b', 't' and 'c' face ids become a more general '1t1', '2b1' and '2t1'.
This lets us use a consistent face naming system regardless of the number of
chips and the number of top and bottom faces on them.

Also add new method of defining layers and faces using layer configuration
files and "example_layer_config.py" to illustrate this.

Co-authored-by: David Janzso <david@meetiqm.com>