Add multigrid and p-element support for Elmer
This commit implements a possibility to use higher order p-elements.
The vectorized `StatElecSolveVec` module is used instead of `StatElecSolve`.
The feature can be used when "multigrid" is used as the linear system solver.

For controlling the feature, there are two new keywords in Elmer
```
export_parameters_elmer={
	'linear_system_method': 'mg', # Default: 'mg'
	'p_element_order': 3,         # Default: 3
	}
```
The first activates the multigrid option and the second sets the
p-element order.

Known issue:
Parallel version gives wrong results for p>3.
This is an issue on Elmer side and is currently being fixed
(EDIT: this should be in order now).

Issue #1094