Centralize elmer mesh refinement
The mesh refinement is based on mesh size fields in both 3D elmer and cross-section elmer. 
Idea is that one can set max mesh element length in the neighbourhood of any entity. The 
entity can be surface or intersection of surfaces. In the future, we might enable entity 
being 3D volume in 3D elmer to set individual mesh refinement on vacuum and substrates. 
But this is not implemented for now.

Update elmer simulation scripts to work with new mesh refinement formatting.

Remove 'algorithm' parameter and use fixed value 5 for now.

Removed `default_workflow` parameter, because it was not the nicest implementation.