PyPI package release
The user may get the stand-alone version of KQCircuits with a simple "pip
install kqcircuits". This is a minimal version that only contains the core KQC
classes. Tests, documentation, scripts, masks, etc. are not included in the
package but they are still supported if the user works from the KQCircuits
source directory.

Documentation generation uses Qt-less rendering making it a lot faster.

Issue #1046