Implement ignore keys for assert dicts
When EPR results are compared in export_and_run tests, E_signal
and E_ground fail because they are almost zero and relative error
can become very large. This issue is solved by implementing a way
to ignore keys that are not applicable.

Extra: implemented a recursion for the dict comparison so that
it is deep.

Issue #1274