Implement Elmer parallel explicit frequency sweep
Number of parallel elmer instances is controlled with workflow parameter
`n_workers`. The same parameter can be used for parallelizing full simulations
in the case of a parameter sweep.

In case of a remote simulation, a similar parameter `n_workers` is used, but within
the sub-dictionary `sbatch_parameters`. Also the logic behind `sbatch_parameters` entries
is modified such that the user requested resources are per simulation while the real
total size of the resource allocation is automatically calculated based on the `n_workers`
and type of simulation

Smaller  changes:
 - Unnecessary `json_` is removed from the suffix of log filenames
 - Documentation updated