Allow placing mask name labels on chips during mask export
This commit makes it possible to add mask name labels to each chip as overlay during mask export, instead of coding the label inside the chip cell. To use this feature, set name_mask="_3" in the chip to add a 4-character placeholder, and set add_mask_name_to_chips=True in MaskSet to enable the label overlay.