Export cells as code: simplify output and support interconnected waveguides
This commit improves the `export_cells_as_code` macro, such that it is possible to connect
waveguides to refpoints defined inline elements in other WaveguideComposite instances. 
For this to work, the inline element must have an instance name defined.

The code generation output is also simplified to make use of more recent features in 
KLayoutView.insert_cell. These changes affect the output when "macro" mode is chosen.