Fix merge_points_and_match_on_edges
Specifically, fix the `fixed_polygon` that is inside the `merge_points_and_match_on_edges` function. The splitting of polygons into small pieces was implemented assuming that i0 and i1 are consecutive instances of the same point in a polygon, but all pairs i0, i1 were iterated using nested loops `for i0 in instances for i1 in instances if i0 != i1`. This can cause an issue when two holes of the same polygon share a point. The implementation is now changed such that only consecutive pairs i0, i1 are iterated.