Fix bug in Ansys geometry import
When loading json in `import_simulation_geometry.py` the order of `layers` dictionary is not conserved. This can cause issue when performing subtraction on layer objects.

The ordering of subtraction operations is now fixed such that each tool layer subtraction is performed before it's used as tool for other subtraction.