Refactor CrossSectionSimulation layers interface on simulation data side
The `CrossSectionSimulation` class implements temporary function that converts the existing data into similar layer format as the `Simulation` class uses, i.e. format of `layers` and `material_dict`. This function will be deleted once the build interface of `CrossSectionSimulation` is refactored.

The Ansys and Elmer cross-section imports are modified to work with new layer formulation in json file. The Ansys cross-section import is refactored to use "excitation" keyword to assign metals as ground or dedicated signals. The Elmer side doesn't yet use the port number information stored in the "excitation" data, but it distinguishes ground and signal using the "excitation" data.

The material_dict definition is written a bit more explicitly in Simulation class.

The excitation numbering is changed so that 0 = ground and 1... = signals. Also floating islands are now signals.