Do not invoke the chip class during mask export documentation step
This commit removes a historical artefact where the chip PCell class was needed to infer the full
set of parameters in the mask export documentation. All parameters are now available in the chip json.

This commit also slightly improves the `skip_extras` flag in mask export: now the chip_json with
parameters is still exported, only the slow operations are skipped.